Safari-609.1.11.1/LayoutTests/fast/loader/cache-encoding-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Invalid character '\u2073'
First load a script with a wrong charset then again with the right one. Second attempt should work and 'scriptSuccess' should be set to true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Loading with the wrong encoding...
PASS Got expected error: 'SyntaxError'
PASS window.scriptSuccess is undefined.
Now loading with the right encoding...
PASS scriptSuccess is true
PASS successfullyParsed is true

TEST COMPLETE
